#update #krypton - level1-2, check for task file existence

+++ krypton/level1/get-password.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-scp -P 2231 krypton1@krypton.labs.overthewire.org:/krypton/krypton1/krypton2 .
+# https://overthewire.org/wargames/krypton/krypton1.html
+
+if [ ! -f krypton2 ]; then
+  scp -P 2231 krypton1@krypton.labs.overthewire.org:/krypton/krypton1/krypton2 .
+fi
 
 PASSWORD=$(cat krypton2 | tr 'A-Za-z' 'N-ZA-Mn-za-m')
